uint64_t is not size_t on OSX (OSX)

+++ builtin/sub_dataset.cc
@@ -530,7 +530,7 @@ struct SubDataset::Itl
         return result;
     }
 
-    virtual size_t getRowCount() const
+    virtual uint64_t getRowCount() const
     {
         return mainChunk.subOutput.size();
     }
